refactor: move sec_event_log decorators into distros. Add tests
