+++ cloudinit/log/security_event_log.py
@@ -14,13 +14,12 @@ Security events are logged in JSON Lines format with standardized fields:
 - event: Event type with optional parameters (e.g., user_created:root,ubuntu)
 - level: INFO, WARN, or CRITICAL
 - description: Human-readable summary
-- host_ip: Optional IP address, included when network information is available.
+- hostname: System hostname
 """
 
-import functools
 import logging
 from enum import Enum
-from typing import Any, Dict, List, Optional, Tuple
+from typing import Any, Dict, List, Optional
 
 from cloudinit import util
 from cloudinit.log import loggers
@@ -54,7 +53,7 @@ class OWASPEventType(Enum):
     # TODO(USER_UPDATED = "user_updated")
 
 
-def _log_security_event(
+def log_security_event(
     event_type: OWASPEventType,
     level: OWASPEventLevel,
     description: str,
@@ -71,14 +70,10 @@ def _log_security_event(
     :param additional_data: Additional context-specific data.
     """
     # cloud-init is the default primary 'actor' for any system change operation
+    params = ["cloud-init"]
     if event_params:
-        event_params.insert(0, "cloud-init")
-    else:
-        event_params = ["cloud-init"]
-
-    event_str = event_type.value
-    if event_params:
-        event_str += ":" + ",".join(event_params)
+        params.extend(event_params)
+    event_str = f"{event_type.value}:{','.join(params)}"
     event = {
         "appid": APP_ID,
         "type": "security",
@@ -89,101 +84,7 @@ def _log_security_event(
     }
     if additional_data:
         # Merge additional non-empty data but don't overwrite core fields
-        for key, value in additional_data.items():
-            if key not in event and value:
-                event[key] = value
-
+        event.update(
+            {k: v for k, v in additional_data.items() if v and k not in event}
+        )
     LOG.log(loggers.SECURITY, event)
-
-
-def sec_log_user_created(func):
-    """A decorator to log a user creation event and group attributes."""
-
-    @functools.wraps(func)
-    def decorator(
-        self, name: str, *args, groups: Optional[List[str]] = None, **kwargs
-    ):
-        if not name:
-            raise RuntimeError(
-                "sec_log_user_created requires positional param name or kwarg"
-            )
-        params = [name]
-        groups_msg = ""
-        if groups is None:
-            groups = []
-        all_groups = groups + self._get_elevated_roles(**kwargs)
-        if all_groups:
-            groups_suffix = ",".join(all_groups)
-            groups_msg = f" in groups: {groups_suffix}"
-            params.append(f"groups:{groups_suffix}")
-
-        response = func(self, name, groups=groups, *args, **kwargs)
-        # User creation operation did not raise an Exception
-        _log_security_event(
-            event_type=OWASPEventType.USER_CREATED,
-            # Treat INFO level as this is prescribed provisioning at launch
-            level=OWASPEventLevel.INFO,
-            description=f"User '{name}' was created{groups_msg}",
-            event_params=params,
-        )
-        return response
-
-    return decorator
-
-
-def sec_log_password_changed_batch(func):
-    @functools.wraps(func)
-    def decorator(self, plist_in: List[Tuple[str, str]], *args, **kwargs):
-        response = func(self, plist_in, *args, **kwargs)
-        for userid, _ in plist_in:
-            _log_security_event(
-                event_type=OWASPEventType.AUTHN_PASSWORD_CHANGE,
-                level=OWASPEventLevel.INFO,
-                description=f"Password changed for user '{userid}'",
-                event_params=[userid],
-            )
-        return response
-
-    return decorator
-
-
-def sec_log_password_changed(func):
-    """A decorator logging a password change event."""
-
-    @functools.wraps(func)
-    def decorator(self, user: str, *args, **kwargs):
-        response = func(self, user, *args, **kwargs)
-        _log_security_event(
-            event_type=OWASPEventType.AUTHN_PASSWORD_CHANGE,
-            level=OWASPEventLevel.INFO,
-            description=f"Password changed for user '{user}'",
-            event_params=[user],
-        )
-        return response
-
-    return decorator
-
-
-def sec_log_system_shutdown(func):
-    """A decorator logging a system shutdown event."""
-
-    @functools.wraps(func)
-    def decorator(cls, mode: str, delay: str, message):
-        if mode == "reboot":
-            event_type = OWASPEventType.SYS_RESTART
-            description = "System restart initiated"
-        else:
-            event_type = OWASPEventType.SYS_SHUTDOWN
-            description = "System shutdown initiated"
-        if message:
-            description += f": {message}"
-
-        _log_security_event(
-            event_type=event_type,
-            level=OWASPEventLevel.INFO,
-            description=description,
-            additional_data={"delay": delay, "mode": mode},
-        )
-        return func(cls, mode=mode, delay=delay, message=message)
-
-    return decorator
